ji.rpkg.template: Your Package Title in Title Case

Your package description. It must end with a period (".") and include relevant bibliographical references if applicable, using the following format: Author et al. (2023) doi:10.5281/zenodo.6619350 .

Author

Maintainer: Pratik Gupte pratik.gupte@lshtm.ac.uk (ORCID)

Other contributors:

Abdul Latif Jameel Institute for Disease and Emergency Analytics [funder]

Imperial College of Science, Technology and Medicine [copyright holder, funder]